feat: redis cache option

+++ src/Cache.php
@@ -11,21 +11,58 @@ namespace hmerritt;
 class Cache
 {
 
-    function __construct()
+    /**
+     * Initiate cache
+     *
+     * @param string $type Type of cache: file | redis
+     * @param array $redisClientOptions Options for redis
+     */
+    function __construct(string $type = "file", array $redisClientOptions = [])
     {
+        $this->type = $type;
+        $this->isRedis = $type === "redis";
+
         /**
-         * Initiate cache database
-         *
+         * @var \Predis\Client
          * @var \Filebase\Database
          */
-        $this->cache = new \Filebase\Database([
-            'dir'            => __DIR__ . DIRECTORY_SEPARATOR . 'cache/films/',
-            'backupLocation' => __DIR__ . DIRECTORY_SEPARATOR . 'cache/films/backups/',
-            'format'         => \Filebase\Format\Json::class,
-            'cache'          => true,
-            'cache_expires'  => 31540000,
-            'pretty'         => false
-        ]);
+        if ($this->isRedis) {
+            $redisOpts = [
+                'scheme'   => 'tcp',
+                'host'     => '127.0.0.1',
+                'port'     => 6379,
+                'password' => '',
+                'database' => 0,
+            ];
+            foreach ($redisClientOptions as $key => $option) {
+                $redisOpts[$key] = $option;
+            }
+
+            $this->redis = new \Predis\Client([
+                'scheme' => $redisOpts['scheme'],
+                'host'   => $redisOpts['host'],
+                'port'   => $redisOpts['port'],
+            ], [
+                'parameters' => [
+                    'password' => $redisOpts['password'],
+                    'database' => $redisOpts['database'],
+                ],
+            ]);
+        } else {
+            $this->cache = new \Filebase\Database([
+                'dir'            => __DIR__ . DIRECTORY_SEPARATOR . 'cache/films/',
+                'backupLocation' => __DIR__ . DIRECTORY_SEPARATOR . 'cache/films/backups/',
+                'format'         => \Filebase\Format\Json::class,
+                'cache'          => true,
+                'cache_expires'  => 31540000,
+                'pretty'         => false
+            ]);
+        }
+    }
+
+    public function isRedis()
+    {
+        return $this->isRedis;
     }
 
     /**
@@ -37,9 +74,13 @@ class Cache
      */
     public function add(string $key, $value)
     {
-        $file = $this->get($key);
-        $file->film = $value;
-        $file->save();
+        if ($this->isRedis) {
+            $this->redis->set($key, json_encode($value));
+        } else {
+            $file = $this->get($key);
+            $file->film = $value;
+            $file->save();
+        }
         return true;
     }
 
@@ -50,8 +91,12 @@ class Cache
      */
     public function delete(string $key): bool
     {
-        $file = $this->get($key);
-        $file->delete();
+        if ($this->isRedis) {
+            $this->redis->del($key);
+        } else {
+            $file = $this->get($key);
+            $file->delete();
+        }
         return true;
     }
 
@@ -63,7 +108,8 @@ class Cache
      */
     public function get(string $key): object
     {
-        return $this->cache->get($key);
+        if ($this->isRedis) return json_decode($this->redis->get($key));
+        else return $this->cache->get($key);
     }
 
     /**
@@ -74,7 +120,7 @@ class Cache
      */
     public function has(string $key): bool
     {
-        return $this->cache->has($key);
+        if ($this->isRedis) return $this->redis->exists($key);
+        else return $this->cache->has($key);
     }
-
 }
